+++ spec/sortah_handler_spec.rb
@@ -11,6 +11,7 @@ describe Sortah::Handler do
       context.should_receive(:lenses).and_return(instance_of(Sortah::Lenses))
       context.should_receive(:routers).and_return(instance_of(Sortah::Routers))
       context.should_receive(:maildir).and_return(instance_of(String))
+      context.should_receive(:error_dest).and_return(instance_of(String))
 
       Sortah::Handler.build_from(context)
     end